1. wrap bpk save button into the bpk card button 2. move the type into index.ts 3. modify README.md
