+++ packages/bpk-component-card-button/index.ts
@@ -15,21 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { render } from '@testing-library/react';
-import { axe } from 'jest-axe';
 
-import BpkSaveButton from './BpkSaveButton';
+import BpkSaveButton, { SIZE_TYPES, STYLE_TYPES } from './src/BpkSaveButton';
 
-describe('BpkSaveButton accessibility tests', () => {
-  it('should not have programmatically-detectable accessibility issues', async () => {
-    const { container } = render(
-      <BpkSaveButton
-        checked={false}
-        accessibilityLabel="Click to save"
-        onCheckedChange={()=>{}}
-      />,
-    );
-    const results = await axe(container);
-    expect(results).toHaveNoViolations();
-  });
-});
+export { BpkSaveButton, SIZE_TYPES, STYLE_TYPES };
